pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepharvest"
version = "1.0.1"
description = "The world's most complete, resilient, multilingual web crawler"
readme = "README.md"
requires-python = ">=3.9"
license = "Apache-2.0"
authors = [
    {name = "DeepHarvest Contributors"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aiohttp>=3.9.0",
    "lxml>=4.9.0",
    "beautifulsoup4>=4.12.0",
    "html5lib>=1.1",
    "playwright>=1.40.0",
    "PyMuPDF>=1.23.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.21",
    "openpyxl>=3.1.0",
    "Pillow>=10.1.0",
    "pytesseract>=0.3.10",
    "extruct>=0.15.0",
    "chardet>=5.2.0",
    "charset-normalizer>=3.3.0",
    "langdetect>=1.0.9",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "simhash>=2.1.0",
    "datasketch>=1.6.0",
    "mmh3>=4.0.0",
    "redis[hiredis]>=5.0.0",
    "boto3>=1.28.0",
    "psycopg2-binary>=2.9.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "prometheus-client>=0.19.0",
    "tqdm>=4.66.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
]
full = [
    "torch>=2.1.0",
    "transformers>=4.35.0",
]

[project.urls]
Homepage = "https://github.com/deepharvest/deepharvest"
Documentation = "https://github.com/deepharvest/deepharvest"
Repository = "https://github.com/deepharvest/deepharvest"
"Bug Tracker" = "https://github.com/deepharvest/deepharvest/issues"

[project.scripts]
deepharvest = "deepharvest.cli.main:cli"

[tool.setuptools]
packages = {find = {}}
include-package-data = true

[tool.setuptools.package-data]
deepharvest = ["py.typed"]

